minor code restructuring
