+++ src/Engine/ProtoAssociative/CodeGen.cs
@@ -2857,7 +2857,7 @@ namespace ProtoAssociative
                     // Handle SSA if the lhs is not an identifier
                     // A non-identifier LHS is any LHS that is not just an identifier name.
                     //  i.e. a[0], a.b, f(), f(x)
-                    if (!ProtoCore.ASTCompilerUtils.IsSingleIdentifier(leftNode))
+                    if (!ASTCompilerUtils.IsSingleIdentifier(leftNode))
                     {
                         EmitSSALHS(ref leftNode, ssaStack, ref astlist);
                     }
