+++ src/Engine/ProtoCore/Utils/CompilerUtils.cs
