+++ src/Engine/ProtoCore/Utils/ParserUtils.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
 using ProtoCore.DSASM;
 using ProtoCore.Lang.Replication;
 using ProtoCore.Utils;
@@ -13,6 +14,65 @@ namespace ProtoCore.Utils
     /// </summary>
     public static class ParserUtils
     {
+        private static string FindIdentifiers(string statement, int equalIndex)
+        {
+
+            // If there is a '{' before the first '=' in the statement,
+            // split the string at '{' and extract the portion after it
+            string identifier = null;
+            int braceIndex = statement.IndexOf('{');
+
+            if (braceIndex != -1)
+            {
+                if (braceIndex < equalIndex)
+                {
+                    statement = statement.Split('{')[1];
+                }
+            }
+
+            identifier = statement.IndexOf('=') == -1 ? null : statement.Split('=')[0];
+
+            return identifier;
+        }
+
+        public static List<string> GetLHSatAssignment(string line, int equalIndex)
+        {
+            var spaceNormalizationRule = new Regex(@"([^\S]+)");
+
+            // the line could have multiple program statements separated by ';'
+            var programStatements = line.Split(';');
+
+            List<string> identifiers = new List<string>();
+            foreach (var statement in programStatements)
+            {
+                if (statement.Equals(string.Empty))
+                    continue;
+
+                string identifier = null;
+                identifier = FindIdentifiers(statement, equalIndex);
+
+                if (identifier == null)
+                    continue;
+                else
+                {
+                    identifier = identifier.Trim();
+
+                    if (identifier.StartsWith(@"//"))
+                        continue;
+                    else if (identifier.Equals("return"))
+                        continue;
+
+                    identifier = spaceNormalizationRule.Replace(identifier, string.Empty);
+
+                    // Check if identifier contains ':' and extract the lhs of the ':'
+                    identifier = identifier.Split(':')[0];
+
+                    identifiers.Add(identifier);
+                }
+            }
+            return identifiers;
+        }
+
         /// <summary>
         /// Retrieves the lhs identifer of a string
         /// </summary>
@@ -52,7 +112,7 @@ namespace ProtoCore.Utils
                 if (index != -1)
                 {
                     string sub;
-                    if (index < input.Length - 2) 
+                    if (index < input.Length - 2)
                     {
                         if (input[index + 1].Equals('\r') && input[index + 2].Equals('\n'))
                         {
@@ -68,5 +128,163 @@ namespace ProtoCore.Utils
             return expr;
         }
 
+        public static string ExtractStatementFromCode(string code, ProtoCore.AST.Node node)
+        {
+            int line = node.line;
+            int col = node.col;
+
+            string stmt = "";
+
+            if (node is ProtoCore.AST.AssociativeAST.IdentifierNode || node is ProtoCore.AST.AssociativeAST.IdentifierListNode
+                || node is ProtoCore.AST.AssociativeAST.FunctionCallNode || node is ProtoCore.AST.AssociativeAST.RangeExprNode
+                || node is ProtoCore.AST.AssociativeAST.StringNode || node is ProtoCore.AST.AssociativeAST.IntNode
+                || node is ProtoCore.AST.AssociativeAST.DoubleNode || node is ProtoCore.AST.AssociativeAST.BooleanNode
+                || node is ProtoCore.AST.AssociativeAST.ExprListNode)
+            {
+                string[] lines = code.Split('\n');
+                int lastLine = lines.Length;
+                int lastCol = lines[lastLine - 1].Length;
+
+                string stmts = ExtractStatementHelper(code, line, col, lastLine, lastCol + 1);
+                string[] induvidualStmts = stmts.Split(';');
+                stmt = induvidualStmts[0] + ";";
+                if (induvidualStmts.Length > 1)
+                {
+                    foreach (char ch in induvidualStmts[1])
+                        if (ch == ' ' || ch == '\n' || ch == '\t')
+                            stmt += ch.ToString();
+                        else
+                            break;
+                }
+                return stmt;
+            }
+
+            stmt = ExtractStatementHelper(code, line, col, node.endLine, node.endCol);
+            return stmt;
+        }
+
+        private static string ExtractStatementHelper(string code, int line, int col, int endLine, int endCol)
+        {
+            int linePos = 1;
+            int charPos = 1;
+            int startIndex = 0;
+            int endIndex = 0;
+            int prevLineLen = 0;
+
+            while (charPos <= code.Length + 1)
+            {
+                if (linePos == line)
+                {
+                    if (charPos == prevLineLen + col)
+                    {
+                        startIndex = charPos;
+                    }
+                }
+                if (linePos == endLine)
+                {
+                    if (charPos == prevLineLen + endCol)
+                    {
+                        if (charPos != code.Length + 1)
+                            endIndex = charPos + 1;
+                        else
+                            endIndex = charPos;
+                        while (charPos < code.Length && code[charPos] == '\n')
+                        {
+                            charPos++;
+                            endIndex++;
+                        }
+                        break;
+                    }
+
+                }
+                if (code[charPos - 1] == '\n')
+                {
+                    prevLineLen = charPos;
+                    ++linePos;
+                }
+
+                ++charPos;
+
+            }
+
+            return code.Substring(startIndex - 1, (endIndex - startIndex));
+        }
+
+        public static List<ProtoCore.AST.Node> GetAstNodes(ProtoCore.AST.Node codeBlockNode)
+        {
+            List<ProtoCore.AST.Node> nodes = new List<ProtoCore.AST.Node>();
+            ProtoCore.AST.AssociativeAST.CodeBlockNode cbn = codeBlockNode as ProtoCore.AST.AssociativeAST.CodeBlockNode;
+            if (cbn != null)
+            {
+                foreach (var n in cbn.Body)
+                {
+                    nodes.Add(n);
+                }
+            }
+            return nodes;
+        }
+
+
+        public static string ExtractCommentStatementFromCode(string code, ProtoCore.AST.AssociativeAST.CommentNode cnode)
+        {
+            string[] lines = code.Split('\n');
+            int lastLine = lines.Length;
+            int lastCol = lines[lastLine - 1].Length;
+            string stmnt = ExtractStatementHelper(code, cnode.line, cnode.col, lastLine, lastCol);
+            
+            int commentLength = cnode.Value.Length;
+            string comment = "";
+            if (stmnt.Length > commentLength)
+            {
+                comment = stmnt.Remove(commentLength);
+                for (int i = commentLength; i < stmnt.Length; i++)
+                {
+                    char ch = stmnt[i];
+                    if (char.IsWhiteSpace(ch))
+                        comment += ch.ToString();
+                    else
+                        break;
+                }
+            }
+            else
+                comment = stmnt;
+            return comment;
+        }
+
+        /// <summary>
+        /// Parses designscript code and returns a ProtoAST CodeBlockNode
+        /// </summary>
+        /// <param name="code"> Source code to parse </param>
+        public static ProtoCore.AST.Node Parse(string code)
+        {
+            Validity.Assert(code != null);
+
+            ProtoCore.Core core = new ProtoCore.Core(new ProtoCore.Options());
+            core.Options.ExecutionMode = ProtoCore.ExecutionMode.Serial;
+            core.ParsingMode = ProtoCore.ParseMode.AllowNonAssignment;
+            core.IsParsingCodeBlockNode = true;
+            core.IsParsingPreloadedAssembly = false;
+
+            return ParseWithCore(code, core);
+        }
+
+        /// <summary>
+        /// Parses desginscript code with specified core and returns a 
+        /// ProtoAST CodeBlockNode
+        /// </summary>
+        /// <param name="code"></param>
+        /// <param name="core"></param>
+        /// <returns></returns>
+        public static ProtoCore.AST.Node ParseWithCore(string code, ProtoCore.Core core)
+        {
+            System.IO.MemoryStream memstream = new System.IO.MemoryStream(System.Text.Encoding.UTF8.GetBytes(code));
+            ProtoCore.DesignScriptParser.Scanner s = new ProtoCore.DesignScriptParser.Scanner(memstream);
+            ProtoCore.DesignScriptParser.Parser p = new ProtoCore.DesignScriptParser.Parser(s, core);
+
+            Validity.Assert(null != p);
+            p.Parse();
+
+            return p.root;
+        }
     }
 }
